+++ lib/components/Modal/ModalBody.tsx
@@ -1,9 +1,10 @@
 import React, { ReactNode } from 'react';
+import styles from './ModalBody.module.scss';
 
 interface IModalBodyProps {
 	children?: ReactNode;
 }
 
 export const ModalBody = ({ children }: IModalBodyProps) => (
-	<div className="modal-body">{children}</div>
+	<div className={styles.modalBody}>{children}</div>
 );